Move shelter form functionality to its own file
